Move NudgeType enum to shared enums
