+++ Bot.Shared/Enums/NudgeType.cs
@@ -1,0 +1,27 @@
+namespace Bot.Shared.Enums;
+
+public enum NudgeType
+{
+    BadPin,
+    TransferFail,
+    BudgetAlert,
+    RecurringCancelled,
+    InvalidBvn,
+    ServiceDown,
+    InvalidNin,
+    WaitingOnMandate,
+    Greeting,
+    Unknown,
+    SignupRequired,
+    AlreadyOnboarded,
+    TimedOut,
+    FloodProtection,
+    Canceled,
+    Help,
+    SignupFailed,
+    BankFailed,
+    KycFailed,
+    BillFailed,
+    RecurringFailed
+}
+
