+++ Bot.Shared/Payloads.cs
@@ -34,29 +34,3 @@ public record SignupPayload(string FullName, string Phone, string NIN, string BV
 public record GreetingPayload(string Message);
 
 public record UnknownPayload(string Message);
-
-// UX
-public enum NudgeType
-{
-    BadPin,
-    TransferFail,
-    BudgetAlert,
-    RecurringCancelled,
-    InvalidBvn,
-    ServiceDown,
-    InvalidNin,
-    WaitingOnMandate,
-    Greeting,
-    Unknown,
-    SignupRequired,
-    AlreadyOnboarded,
-    TimedOut,
-    FloodProtection,
-    Canceled,
-    Help,
-    SignupFailed,
-    BankFailed,
-    KycFailed,
-    BillFailed,
-    RecurringFailed
-}
